+++ pcap4j-packetfactory-static/src/main/java/org/pcap4j/packet/factory/StaticIpV6RoutingDataFactory.java
@@ -14,23 +14,23 @@ import org.pcap4j.packet.IllegalRawDataException;
 import org.pcap4j.packet.IpV6ExtRoutingPacket.IpV6RoutingData;
 import org.pcap4j.packet.IpV6RoutingSourceRouteData;
 import org.pcap4j.packet.UnknownIpV6RoutingData;
-import org.pcap4j.packet.namednumber.IpV6RoutingHeaderType;
+import org.pcap4j.packet.namednumber.IpV6RoutingType;
 
 /**
  * @author Kaito Yamada
  * @since pcap4j 0.9.16
  */
 public final class StaticIpV6RoutingDataFactory
-implements PacketFactory<IpV6RoutingData, IpV6RoutingHeaderType> {
+implements PacketFactory<IpV6RoutingData, IpV6RoutingType> {
 
   private static final StaticIpV6RoutingDataFactory INSTANCE
     = new StaticIpV6RoutingDataFactory();
-  private final Map<IpV6RoutingHeaderType, Instantiater> instantiaters
-    = new HashMap<IpV6RoutingHeaderType, Instantiater>();
+  private final Map<IpV6RoutingType, Instantiater> instantiaters
+    = new HashMap<IpV6RoutingType, Instantiater>();
 
   private StaticIpV6RoutingDataFactory() {
     instantiaters.put(
-      IpV6RoutingHeaderType.SOURCE_ROUTE, new Instantiater() {
+      IpV6RoutingType.SOURCE_ROUTE, new Instantiater() {
         @Override
         public IpV6RoutingData newInstance(
           byte[] rawData, int offset, int length
@@ -55,7 +55,7 @@ implements PacketFactory<IpV6RoutingData, IpV6RoutingHeaderType> {
 
   @Override
   public IpV6RoutingData newInstance(
-    byte[] rawData, int offset, int length, IpV6RoutingHeaderType number
+    byte[] rawData, int offset, int length, IpV6RoutingType number
   ) {
     if (rawData == null || number == null) {
       StringBuilder sb = new StringBuilder(40);
@@ -88,7 +88,7 @@ implements PacketFactory<IpV6RoutingData, IpV6RoutingHeaderType> {
   }
 
   @Override
-  public Class<? extends IpV6RoutingData> getTargetClass(IpV6RoutingHeaderType number) {
+  public Class<? extends IpV6RoutingData> getTargetClass(IpV6RoutingType number) {
     if (number == null) {
       throw new NullPointerException("number must not be null.");
     }